+++ src/interactions/commands/utility/pokedex.js
@@ -1,58 +1,31 @@
-const { Command } = require("@src/structures");
-const { MessageEmbed, Message, CommandInteraction, CommandInteractionOptionResolver } = require("discord.js");
+const { SlashCommand } = require("@src/structures");
+const { MessageEmbed, CommandInteraction } = require("discord.js");
 const { EMOJIS, MESSAGES, EMBED_COLORS } = require("@root/config.js");
 const { getJson } = require("@utils/httpUtils");
 const outdent = require("outdent");
 
-module.exports = class Pokedex extends Command {
+module.exports = class Pokedex extends SlashCommand {
   constructor(client) {
     super(client, {
       name: "pokedex",
       description: "shows pokemon information",
-      cooldown: 5,
-      command: {
-        enabled: true,
-        usage: "<pokemon>",
-        minArgsCount: 1,
-        category: "UTILITY",
-        botPermissions: ["EMBED_LINKS"],
-      },
-      slashCommand: {
-        enabled: true,
-        options: [
-          {
-            name: "pokemon",
-            description: "pokemon name to get information for",
-            type: "STRING",
-            required: true,
-          },
-        ],
-      },
+      enabled: true,
+      options: [
+        {
+          name: "pokemon",
+          description: "pokemon name to get information for",
+          type: "STRING",
+          required: true,
+        },
+      ],
     });
   }
 
   /**
-   * @param {Message} message
-   * @param {string[]} args
-   */
-  async messageRun(message, args) {
-    const response = await getJson(`https://pokeapi.glitch.me/v1/pokemon/${args}`);
-    if (response.status === 404) return message.reply("```The given pokemon is not found```");
-    if (!response.success) return message.reply(MESSAGES.API_ERROR);
-
-    const json = response.data[0];
-    if (!json) return;
-    const embed = buildEmbed(json);
-
-    message.channel.send({ embeds: [embed] });
-  }
-
-  /**
    * @param {CommandInteraction} interaction
-   * @param {CommandInteractionOptionResolver} options
    */
-  async interactionRun(interaction, options) {
-    const pokemon = options.getString("pokemon");
+  async run(interaction) {
+    const pokemon = interaction.options.getString("pokemon");
 
     const response = await getJson(`https://pokeapi.glitch.me/v1/pokemon/${pokemon}`);
     if (response.status === 404) return interaction.followUp("```The given pokemon is not found```");
@@ -62,7 +35,7 @@ module.exports = class Pokedex extends Command {
     if (!json) return;
     const embed = buildEmbed(json);
 
-    interaction.followUp({ embeds: [embed] });
+    return interaction.followUp({ embeds: [embed] });
   }
 };
